Test access to the file for better security
The problem is that user could query whether a document existed in girder
even if they don't have access to it. The user would try to download it
by its hash and if he received a 403, they knew the document existed.

To prevent this, we can manually go through the file found with the given
hash and test for permission. The first file that the user has access to
is downloaded.